refactor(errors): extract error classification into errors/ module
Spec 2.3 complete: Extract error classification functions from
provider-utils.ts into dedicated errors/ module.

Create errors/classification.ts with:
- toErrorMessage, extractErrorMessage, truncateForLog
- messageIndicatesQuotaError, payloadLooksLikeError
- responseIsEventStream, streamPayloadIndicatesQuotaError
- responseIndicatesQuotaError, responseIndicatesMissingModel
- responseIndicatesModelNotSupportedForAccount, summarizeUpstreamError
- QUOTA_ERROR_PATTERNS, MODEL_NOT_SUPPORTED_WITH_CHATGPT_PATTERNS
- UpstreamErrorSummary interface

Update imports across codebase:
- app.ts, routes/*.ts
- provider-strategy/*.ts
- federation/*.ts
- token-refresh-handlers.ts

Remove functions from provider-utils.ts.

All 579 tests pass.
