+++ src/lib/federation/bridge-fallback.ts
@@ -7,7 +7,7 @@ import {
 } from "../request-utils.js";
 import {
   toErrorMessage,
-} from "../provider-utils.js";
+} from "../errors/index.js";
 import {
   bridgeCapabilitySupportsPath,
   bridgeCapabilitySupportsModel,
