+++ src/lib/federation/federated-fallback.ts
@@ -5,7 +5,7 @@ import type { FastifyReply, FastifyInstance } from "fastify";
 import { buildForwardHeaders } from "../proxy.js";
 import { normalizeRequestedModel } from "../request-utils.js";
 import { fetchWithResponseTimeout } from "../http/index.js";
-import { toErrorMessage } from "../provider-utils.js";
+import { toErrorMessage } from "../errors/index.js";
 import {
   shareModeAllowsRelay,
   shareModeAllowsWarmImport,
