+++ src/lib/provider-strategy/strategies/ollama-cloud.ts
@@ -5,7 +5,8 @@ import {
   ollamaToChatCompletion,
   streamOllamaNdjsonToChatCompletionSse,
 } from "../../ollama-compat.js";
-import { sendOpenAiError, toErrorMessage } from "../../provider-utils.js";
+import { sendOpenAiError } from "../../provider-utils.js";
+import { toErrorMessage } from "../../errors/index.js";
 import { BaseProviderStrategy } from "../base.js";
 import {
   buildPayloadResult,
